style: responsive QR Code screen
